pyproject.toml
[tool.poetry]
name = "etl-dbt-movie-challenge"
version = "0.1.0"
description = ""
authors = ["Santiago Orozco <santiago.orozco@prediktia.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
snowflake-connector-python = "^3.10.0"
pandas = "^2.2.2"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.3"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
